`LossyDictionary`: added failing test and fix for key conversion issue
The property wrapper `LossyDictionary` incorrectly decodes keys with `KeyDecodingStrategy.convertFromSnakeCase`.
The test simply encodes and decodes a sample data, to illustrate that the conversion isn't idempotent.

Keys were being converted to camel case, which is inconsistent from the default behavior that `container.decode([String: Value].self, forKey: key)` would have.
The reason for this is some implementation detail in `Foundation` that ignores the `keyDecodingStrategy` when decoding "raw" dictionaries versus property names.

To deal with this, this decodes the strings first as "raw", and then matches them to the corresponding converted key.